backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.2.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1
celery==5.3.4
librosa==0.10.1
numpy==1.26.0
scipy==1.11.4
pydantic==2.7.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
python-dotenv==1.0.0
Pillow==10.1.0
soundfile==0.12.1
pydub==0.25.1
setuptools==68.0.0